server/requirements.txt
# server/requirements.txt
Flask
Flask-Cors
Flask-Sock
pymongo
python-dotenv
deepgram-sdk
websockets